Move context-menu init code (#553)
